pixie/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixie-api"
version = "0.1.0"
description = "Pixie API service"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Guru", email = "gp3.singh@gmail.com"}
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=0.19.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "mcp-ui-server",
    "supabase>=2.0.0",
    "psycopg[binary,pool]>=3.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "isort>=5.0",
    "mypy>=0.900",
    "flake8>=3.9"
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=app --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
